# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: ERP Supplier Management API
  description: Production-ready Supplier Management backend for ERP system.
  version: "1.0.0"
servers:
  - url: /
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────────
  # Reusable Parameters
  # ─────────────────────────────────────────
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
        example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: "id"
        enum: [id, en]
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    StatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/SupplierStatus"

  # ─────────────────────────────────────────
  # Reusable Responses
  # ─────────────────────────────────────────
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SupplierStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/SupplierStats"

    # ── Supplier List ──────────────────────
...
    SupplierListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/SupplierListItem"
        meta:
          $ref: "#/components/schemas/Meta"

    # ── Supplier Detail ────────────────────
...
paths:
...
  /suppliers/stats:
    get:
      tags: ["Supplier: Management"]
      summary: Get supplier dashboard statistics (Total, New, Avg Cost, Blocked)
      operationId: GetSupplierStats
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Stats retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # SUPPLIER LIST & CREATE
  # ─────────────────────────────────────────
  /suppliers:
    get:
      tags: ["Supplier: Management"]
      summary: List suppliers with pagination and filters
      operationId: ListSuppliers
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/StatusFilter"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Supplier list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: ["Supplier: Management"]
      summary: Create new supplier (from New Supplier modal)
      operationId: CreateSupplier
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # SUPPLIER DETAIL, UPDATE, DELETE
  # ─────────────────────────────────────────
  /suppliers/{supplierID}:
    parameters:
      - $ref: "#/components/parameters/SupplierID"
      - $ref: "#/components/parameters/AcceptLanguage"

    get:
      tags: ["Supplier: Management"]
      summary: Get supplier detail (Overview tab)
      operationId: GetSupplier
      parameters:
        - $ref: "#/components/parameters/Authorization"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Supplier detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: ["Supplier: Management"]
      summary: Update supplier basic info
      operationId: UpdateSupplier
      parameters:
        - $ref: "#/components/parameters/Authorization"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: ["Supplier: Management"]
      summary: Soft delete supplier
      operationId: DeleteSupplier
      parameters:
        - $ref: "#/components/parameters/Authorization"
      security:
        - BearerAuth: []
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # BLOCK / UNBLOCK (Block/Unblock button)
  # ─────────────────────────────────────────
  /suppliers/{supplierID}/block:
    post:
      tags: ["Supplier: Management"]
      summary: Block or unblock a supplier
      operationId: BlockSupplier
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockSupplierRequest"
      responses:
        "200":
          description: Block status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockSupplierResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # STAGE WORKFLOW (Next Stage button)
  # ─────────────────────────────────────────
  /suppliers/{supplierID}/next-stage:
    post:
      tags: ["Supplier: Stage"]
      summary: Advance supplier to next workflow stage (Draft→Review→Assessment→Active)
      operationId: AdvanceSupplierStage
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceStageRequest"
      responses:
        "200":
          description: Stage advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /suppliers/{supplierID}/stage-history:
    get:
      tags: ["Supplier: Stage"]
      summary: Get stage transition history
      operationId: GetStageHistory
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Stage history list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageHistoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # ADDRESS TAB (multiple addresses)
  # ─────────────────────────────────────────
  /suppliers/{supplierID}/addresses:
    get:
      tags: ["Supplier: Addresses"]
      summary: Get all addresses for a supplier (Address tab)
      operationId: GetSupplierAddresses
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Address list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: ["Supplier: Addresses"]
      summary: Add address to supplier
      operationId: AddSupplierAddress
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
      responses:
        "201":
          description: Address added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /suppliers/{supplierID}/addresses/{addressID}:
    parameters:
      - $ref: "#/components/parameters/SupplierID"
      - $ref: "#/components/parameters/AddressID"
      - $ref: "#/components/parameters/AcceptLanguage"

    put:
      tags: ["Supplier: Addresses"]
      summary: Update a supplier address
      operationId: UpdateSupplierAddress
      parameters:
        - $ref: "#/components/parameters/Authorization"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
      responses:
        "200":
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: ["Supplier: Addresses"]
      summary: Delete a supplier address
      operationId: DeleteSupplierAddress
      parameters:
        - $ref: "#/components/parameters/Authorization"
      security:
        - BearerAuth: []
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /suppliers/{supplierID}/addresses/{addressID}/set-main:
    post:
      tags: ["Supplier: Addresses"]
      summary: Set an address as main (radio button)
      operationId: SetMainSupplierAddress
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/AddressID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Main address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # CONTACTS TAB
  # ─────────────────────────────────────────
...
  /suppliers/{supplierID}/contacts/{contactID}/set-primary:
    post:
      tags: ["Supplier: Contacts"]
      summary: Set a contact as primary (Main radio button)
      operationId: SetPrimarySupplierContact
      parameters:
        - $ref: "#/components/parameters/SupplierID"
        - $ref: "#/components/parameters/ContactID"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/AcceptLanguage"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Primary contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ─────────────────────────────────────────
  # GROUPS TAB
  # ─────────────────────────────────────────
...